// frontend-admin/src/api/generated/model/rksvReminderRegisterStatusItemDto.ts
/**
 * Generated by orval v6.31.0 🍺
 * Do not edit manually.
 * Kasse API
 * Registrierkasse API - RKSV uyumlu kasa sistemi
 * OpenAPI spec version: v1
 */
import type { RksvReminderStatusDto } from './rksvReminderStatusDto';

export interface RksvReminderRegisterStatusItemDto {
  cashRegisterId?: string;
  status?: RksvReminderStatusDto;
}